requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
filelock==3.9.0
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
snscrape==0.6.0.20230303
soupsieve==2.4
textblob==0.17.1
tqdm==4.65.0
urllib3==1.26.14
Werkzeug==2.2.3
